% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imcol.R
\name{calcIndices}
\alias{calcIndices}
\title{Title}
\usage{
calcIndices(collection, indices)
}
\arguments{
\item{collection}{an rgee image collection}

\item{indices}{A list of indices to be calculated. Possible indices are 'NDVI', 'NDWI', 'EVI', 'NDSI'.}
}
\value{
a collection of images with the calculated indices
}
\description{
Title
}
\examples{
calcIndices(collection, c('NDVI', 'NDWI'))
}
