% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imcol.R
\name{defCol}
\alias{defCol}
\title{Title}
\usage{
defCol(
  traj,
  product,
  bands = c("blue", "green", "red", "NIR", "SWIR"),
  cloudcover
)
}
\arguments{
\item{product}{one of 'S2' or 'L8'}

\item{bands}{a list of bands to be downloaded}

\item{cloudcover}{the maximum cloud cover percentage as number}

\item{movefile}{a move file}
}
\value{
a collection of images in rgee format
}
\description{
Title
}
\examples{
defCol(movefile, 'S2', c('B2', 'B3', 'B4'), 10)

}
