% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imcol.R
\name{downRast}
\alias{downRast}
\title{Title}
\usage{
downRast(traj, collection, filename = "img", median = TRUE, scale = 30)
}
\arguments{
\item{collection}{an rgee image collection}

\item{filename}{the name of the file to be downloaded}

\item{median}{a boolean to decide if the median image should be downloaded}

\item{scale}{the scale of the image in meters}

\item{traj_sf}{a trajectory in sf format}
}
\value{
download the images to the local machine
}
\description{
Title
}
\examples{
downRast(traj_sf, collection, 'img', TRUE, 10)
}
