% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrate_move.R
\name{intergrate_move}
\alias{intergrate_move}
\title{Title}
\usage{
intergrate_move(move_data)
}
\arguments{
\item{move_data}{A move data frame}
}
\value{
An sf object containing the information of the move data, possibly temporally resampled
}
\description{
Title
}
\examples{
integrate_move(move_data)
}
