% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jakobs_functions.R
\name{long2UTM}
\alias{long2UTM}
\title{Title}
\usage{
long2UTM(long)
}
\arguments{
\item{long}{A numeric value representing the longitude}
}
\value{
A numeric value representing the UTM zone
}
\description{
Title
}
\examples{
long2UTM(long)
}
