% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reclass.R
\name{rast.UD}
\alias{rast.UD}
\title{Title}
\usage{
rast.UD(x, DF = "CDF", ...)
}
\arguments{
\item{x}{occurence data - UD object}

\item{DF}{Rasterize the probability density function "PDF", probability mass function "PMF", or cumulative distribution function "CDF".}

\item{...}{}
}
\value{
A terra object
}
\description{
Title
}
\examples{
rast(occurence_data)
}
