% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reclass.R
\name{reclassNA}
\alias{reclassNA}
\title{Title}
\usage{
reclassNA(rast, maxNA)
}
\arguments{
\item{rast}{terra object}

\item{maxNA}{value below which values are to be set to NA}
}
\value{
terra object with values below maxNA set to NA
}
\description{
Title
}
\examples{
rast <- reclassNA(rast, 0.05)
}
