% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date_time_handling.R
\name{resampleYears}
\alias{resampleYears}
\title{Title}
\usage{
resampleYears(traj)
}
\arguments{
\item{traj}{a move object}
}
\value{
a resampled move object
}
\description{
Title
}
\examples{
resampleYears(traj)
}
