% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ctmm_handling.R
\name{traj_mod}
\alias{traj_mod}
\title{Title}
\usage{
traj_mod(traj)
}
\arguments{
\item{traj}{A move file}
}
\value{
A list with the following ctmm elements: svf, guess, fits, res, akde, od, sim
}
\description{
Title
}
\examples{
traj_model <- traj_mod(traj)
}
