% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visCtmm_plot.R
\name{visOcc_map}
\alias{visOcc_map}
\title{Title}
\usage{
visOcc_map(traj_model, img_filename, info)
}
\arguments{
\item{traj_model}{a ctmm model, output of traj_mod()}

\item{img_filename}{the name of the image file}
}
\value{
a map of the akde on top of a satellite image
}
\description{
Title
}
\examples{
visOcc_AKDE_map(traj_model, 'img_filename')
}
