% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visCtmm_plot.R
\name{visOcc_plot}
\alias{visOcc_plot}
\title{Title}
\usage{
visOcc_plot(traj_model, img_filename, band, info, plottype = "ggplot")
}
\arguments{
\item{traj_model}{a ctmm model, output of traj_mod()}

\item{img_filename}{the name of the image file}

\item{band}{band name to be plotted}

\item{info}{'akde' or 'od'}

\item{plottype}{'ggplot' or 'raster'}
}
\value{
a plot of the occurence vs the band
}
\description{
Title
}
\examples{
visOcc_plot(traj_model, 'img_filename', 'NDVI', 'akde', 'ggplot')
}
