% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visCtmm_plot.R
\name{visOcc_UD_plot}
\alias{visOcc_UD_plot}
\title{Title}
\usage{
visOcc_UD_plot(traj_model, img_filename, band)
}
\arguments{
\item{traj_model}{a ctmm model, output of traj_mod()}

\item{img_filename}{the name of the image file}

\item{band}{the band to be plotted}
}
\value{
a plot of the occurence vs the band
}
\description{
Title
}
\examples{
visOcc_UD_plot(traj_model, 'img_filename', 'B4')
}
