% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vis.R
\name{visTrajDown}
\alias{visTrajDown}
\title{Title}
\usage{
visTrajDown(traj, filename)
}
\arguments{
\item{traj}{a move object of a trajectory}

\item{filename}{the filename of the downloaded raster}
}
\value{
a map with the trajectory points and the images using ggplot
}
\description{
Title
}
\examples{
visTrajDown(traj, 'img')
}
