% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vis.R
\name{visTrajEe}
\alias{visTrajEe}
\title{Title}
\usage{
visTrajEe(traj, collection, bands)
}
\arguments{
\item{collection}{an rgee image collection}

\item{movefile}{a move file}
}
\value{
a map with the trajectory points and the images
}
\description{
Title
}
\examples{
visTraj(movefile, collection)

}
